requirements.txt
pillow==8.3.2
numpy==1.23.5
joblib==1.1.0
